backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
supabase>=2.27.1
httpx>=0.26.0
python-dotenv==1.0.0
psutil==7.1.3
